refactor and examples
